Move props getter to their own service
