+++ app/services/props/getBoxProps.ts
@@ -1,5 +1,5 @@
 import { BoxPropsSchema } from "~/components/Box";
-import type { StrapiBox } from "../models/StrapiBox";
+import type { StrapiBox } from "../cms/models/StrapiBox";
 import { omitNull } from "~/util/omitNull";
 
 export const getBoxProps = (cmsData: StrapiBox) =>
