+++ app/services/props/getHeaderProps.ts
@@ -1,5 +1,5 @@
 import { HeaderPropsSchema } from "~/components/Header";
-import type { StrapiHeader } from "../models/StrapiHeader";
+import type { StrapiHeader } from "../cms/models/StrapiHeader";
 import { omitNull } from "~/util/omitNull";
 
 export const getHeaderProps = (cmsData: StrapiHeader) => {
