+++ app/services/props/getHeadingProps.ts
@@ -1,5 +1,5 @@
 import { HeadingPropsSchema } from "~/components/Heading";
-import type { StrapiHeading } from "../models/StrapiHeading";
+import type { StrapiHeading } from "../cms/models/StrapiHeading";
 import { omitNull } from "~/util/omitNull";
 
 export const getHeadingProps = (cmsData: StrapiHeading) => {
