+++ app/services/props/getImageProps.ts
@@ -1,6 +1,6 @@
 import { ImagePropsSchema } from "~/components/Image";
 import { omitNull } from "~/util/omitNull";
-import type { StrapiImage } from "../models/StrapiImage";
+import type { StrapiImage } from "../cms/models/StrapiImage";
 
 export const getImageProps = (cmsData: StrapiImage) => {
   if (!cmsData.data) return undefined;
